Change Progress, DownSpeed, and UpSpeed properties to use double type for improved precision

+++ Domain/DTOs/DownloadItem.cs
@@ -4,9 +4,9 @@ public record DownloadItem : SearchResult
 {
     public required string SavePath { get; init; }
     public required DownloadStatus Status { get; init; }
-    public required int Progress { get; init; }
-    public required long DownSpeed { get; init; }
-    public required long UpSpeed { get; init; }
+    public required double Progress { get; init; }
+    public required double DownSpeed { get; init; }
+    public required double UpSpeed { get; init; }
     public required long Eta { get; init; }
 }
